fix bug in initialization, determination of equipment present
Currently EquipmentInfo.infoForKey("EQ_XXX") had been used to determine,
if an equipment is present. This is wrong, it justs reveals, if the
equipment is known at all to Oolite. To find out whether it attached to
the ship, we have to use player.ship.equipmentStatus. This now has been
fixed for the initialization process.